Refactor to add BaseMultimediaFileManager
